PHPStan L7

+++ src/MyMetaDate.php
@@ -59,6 +59,6 @@ class MyMetaDate extends MyMetaField
 
     public function displayValue(string $value): string
     {
-        return (string) date('Y-m-d H:i', strtotime($value)) . ' UTC';
+        return (string) date('Y-m-d H:i', (int) strtotime($value)) . ' UTC';
     }
 }
